Feature/aq 32 users list

+++ src/pages/Users.tsx
@@ -1,14 +1,8 @@
-/**
- * Página de Usuarios - Acuaterra (Users).
- * Visual: Basado en el diseño de Figma,
- * La lógica de CRUD se mantiene intacta.
- */
-
 import type { FunctionComponent } from "react";
 // eslint-disable-next-line no-duplicate-imports
 import { useState } from "react";
-import type { UserRequest, User } from "../common/types";
-import UserTable from "../components/ui/table/table";
+import type { UserRequestV2, UserResponse } from "../common/types";
+import TableWithActions from "../components/ui/table/tableWithActions";
 import useUsers from "../hooks/useUsers";
 import RegisterUserModal from "../components/ui/modals/registerUserModal";
 import UpdateUserModal from "../components/ui/modals/updateUserModalProps";
@@ -16,17 +10,13 @@ import useRegisterUser from "../hooks/useRegisterUser";
 import { deleteUser, updateUser } from "../services/userService";
 import Layout from "../components/layout/layout";
 import { useNavigate } from "@tanstack/react-router";
-
-// Íconos e imágenes
 import closeSessionIcon from "../assets/images/cerrar-sesion.png";
 import userIcon from "../assets/images/userlogo.png";
 import moduleIcon from "../assets/images/module.png";
 import homeIcon from "../assets/images/home.png";
 import acuaterraLogo from "../assets/images/logo.png";
 import reportIcon from "../assets/images/reporte.png";
-import binnacleIcon from "../assets/images/bitacora.png";
-
-// Toast y Spinner
+import fishIcon from "../assets/images/pez.png";
 import Toast from "../components/ui/Toast";
 import Spinner from "../components/ui/Spinner";
 
@@ -35,62 +25,53 @@ export const Users: FunctionComponent = () => {
   const [page, setPage] = useState(1);
   const [showModal, setShowModal] = useState(false);
   const [showUpdateModal, setShowUpdateModal] = useState(false);
-  const [selectedUser, setSelectedUser] = useState<User | null>(null);
+  const [selectedUser, setSelectedUser] = useState<UserResponse | null>(null);
   const [reload, setReload] = useState(false);
   const pageSize = 10;
-  const { users, loading, error } = useUsers(page, pageSize, reload);
+  const { users, loading, error, total } = useUsers(page, pageSize, reload);
   const { registerUser } = useRegisterUser();
-
-  // Para mostrar un toast cuando se registre un usuario exitosamente
   const [showToast, setShowToast] = useState(false);
 
-  const handleRegisterUser = async (userData: UserRequest): Promise<void> => {
+  const handleRegisterUser = async (userData: UserRequestV2): Promise<void> => {
     try {
       await registerUser(userData);
       setReload(!reload);
       setShowModal(false);
-      setShowToast(true); // Mostramos la notificación
+      setShowToast(true);
     } catch (error) {
       console.error(error);
     }
   };
 
   const handleDeleteUser = async (userId: number): Promise<void> => {
-    await deleteUser(userId);
-    setReload(!reload);
+    const confirmed = window.confirm("¿Estás seguro de que deseas eliminar este usuario?");
+    if (confirmed) {
+      await deleteUser(userId);
+      setReload(!reload);
+    }
   };
 
-  const handleUpdateUser = async (                                          // Merged Const
-    userId: number,
-    userData: UserRequest
-  ): Promise<void> => {
+  const handleUpdateUser = async (userId: number, userData: UserRequestV2): Promise<void> => {
     await updateUser(userId, userData);
     setReload(!reload);
   };
 
-  const handleOpenUpdateModal = (user: User): void => {                     // Merged Const
+  const handleOpenUpdateModal = (user: UserResponse): void => {
     setSelectedUser(user);
     setShowUpdateModal(true);
   };
 
-
-
-  // Integración de la página de usuarios
-
   return (
     <Layout>
-      <div className="flex min-h-screen bg-white">
-        {/* Sidebar con fondo gris (bg-gray-300) */}
-        <aside className="w-64 bg-gray-300 border-r border-gray-300 flex flex-col">
+      <div className="flex flex-col md:flex-row min-h-screen bg-white">
+        {/* Sidebar */}
+        <aside className="w-full md:w-64 bg-gray-300 border-r border-gray-300 flex flex-col">
           <div className="p-4 flex flex-col items-center">
             <img alt="Acuaterra Logo" className="h-16 mb-2" src={acuaterraLogo} />
             <p className="text-gray-700 font-semibold">Bienvenido, usuario!</p>
           </div>
           <nav className="flex-1">
-
-            
-            {/* Grupo 1: "Inicio", "Usuarios" y "Módulos" */}
-            <ul className="space-y-20 mt-20">
+            <ul className="space-y-4 md:space-y-20 mt-4 md:mt-20">
               <li
                 className="flex items-center p-2 cursor-pointer transition-all duration-300 hover:bg-gray-300 hover:scale-105"
                 onClick={() => navigate({ to: "/newHome" })}
@@ -98,7 +79,14 @@ export const Users: FunctionComponent = () => {
                 <img alt="Inicio" className="h-6 w-6 mr-2" src={homeIcon} />
                 <span className="font-bold">Inicio</span>
               </li>
-              <li className="flex items-center p-2 bg-gray-300 transition">
+              <li
+                className="flex items-center p-2 cursor-pointer transition-all duration-300 hover:bg-gray-400 hover:scale-105"
+                onClick={() => navigate({ to: "/farm" })}
+              >
+                <img alt="Granjas" className="h-6 w-6 mr-2" src={moduleIcon} />
+                <span className="font-bold">Granjas</span>
+              </li>
+              <li className="flex items-center p-2 cursor-pointer transition-all duration-300 hover:bg-gray-400 hover:scale-105 bg-gray-400 text-white border-2 border-gray-400 rounded-lg">
                 <img alt="Usuarios" className="h-6 w-6 mr-2" src={userIcon} />
                 <span className="font-bold">Usuarios</span>
               </li>
@@ -106,7 +94,7 @@ export const Users: FunctionComponent = () => {
                 className="flex items-center p-2 cursor-pointer transition-all duration-300 hover:bg-gray-300 hover:scale-105"
                 onClick={() => navigate({ to: "/module" })}
               >
-                <img alt="Módulos" className="h-6 w-6 mr-2" src={moduleIcon} />
+                <img alt="Módulos" className="h-6 w-6 mr-2" src={fishIcon} />
                 <span className="font-bold">Módulos</span>
               </li>
               <li
@@ -116,17 +104,10 @@ export const Users: FunctionComponent = () => {
                 <img alt="Reporte" className="h-6 w-6 mr-2" src={reportIcon} />
                 <span className="font-bold">Reporte</span>
               </li>
-              <li
-              className="flex items-center p-2 cursor-pointer transition-all duration-300 hover:bg-gray-300 hover:scale-105"
-              onClick={() => navigate({ to: "/bitacoras" })}
-                >
-              <img alt="Reporte" className="h-6 w-6 mr-2" src={binnacleIcon} />
-              <span className="font-bold">Bitacoras</span>
-             </li>
-
             </ul>
-            {/* Grupo 2: "Cerrar Sesión" en un bloque separado */}
-            <div className="mt-60">
+
+            
+            <div className="mt-4 md:mt-20">
               <ul className="space-y-4">
                 <li
                   className="flex items-center p-2 cursor-pointer transition-all duration-300 hover:bg-gray-300 hover:scale-105"
@@ -138,7 +119,7 @@ export const Users: FunctionComponent = () => {
               </ul>
             </div>
           </nav>
-          {/* Footer: Texto del footer subido un poco */}
+
           <div className="p-0">
             <p className="text-center text-xs mt-2">
               versión 1.0 <br />
@@ -147,15 +128,15 @@ export const Users: FunctionComponent = () => {
           </div>
         </aside>
 
-        {/* Contenido principal */}
-        <main className="flex-1 p-6">
-          <p className="text-gray-700 font-semibold">Lista de Usuarios</p>
-          {loading && (
-            <div className="mt-4">
+       
+        <main className="flex-1 p-6 bg-white">
+          <h1 className="text-2xl font-bold mb-4">Lista de Usuarios</h1>
+
+          {loading ? (
+            <div className="flex justify-center items-center h-32">
               <Spinner />
             </div>
-          )}
-          {error && (
+          ) : error ? (
             <div className="mt-4 text-red-500 flex items-center">
               <svg className="w-5 h-5 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                 <path
@@ -167,59 +148,35 @@ export const Users: FunctionComponent = () => {
               </svg>
               <p>Error: {String(error)}</p>
             </div>
-          )}
-          {!loading && !error && (
-            <div className="overflow-x-auto">
-              <UserTable
-                users={users}
-                onDeleteUser={handleDeleteUser}
-                onUpdateUser={handleOpenUpdateModal}
+          ) : (
+            <div className="border border-gray-300 rounded-lg p-4 shadow-md">
+              <TableWithActions
+                data={users}
+                error={error}
+                limit={pageSize}
+                loading={loading}
+                page={page}
+                setLimit={() => {}}
+                setPage={setPage}
+                total={total}
+                columns={[
+                  { header: "ID", accessor: "id" },
+                  { header: "Name", accessor: "name" },
+                  { header: "Email", accessor: "email" },
+                  { header: "DNI", accessor: "dni" },
+                  { header: "Role", accessor: "rol", render: (user: UserResponse) => user.rol.name },
+                  { header: "Address", accessor: "address" },
+                  { header: "Created At", accessor: "createdAt" },
+                  { header: "Updated At", accessor: "updatedAt" },
+                ]}
+                onAdd={() => { setShowModal(true); }}
+                onDelete={handleDeleteUser}
+                onEdit={handleOpenUpdateModal}
               />
             </div>
           )}
-          <div className="flex justify-between mt-4">
-            <button
-              disabled={page === 1}
-              className="
-                bg-blue-600 hover:bg-blue-700 text-white font-semibold px-4 py-2 
-                rounded transition focus:outline-none focus:ring-2 focus:ring-blue-300
-              "
-              onClick={() => {
-                setPage(page - 1);
-              }}
-            >
-              Previous
-            </button>
-            <div className="flex flex-col items-center mt-4">
-              <button
-                className="
-                  mb-4 p-2 bg-green-600 hover:bg-green-700 text-white font-semibold 
-                  rounded transition focus:outline-none focus:ring-2 focus:ring-green-300
-                "
-                onClick={() => {
-                  setShowModal(true);
-                }}
-              >
-                Register User
-              </button>
-            </div>
-            <button
-              className="
-                bg-blue-600 hover:bg-blue-700 text-white font-semibold px-4 py-2 
-                rounded transition focus:outline-none focus:ring-2 focus:ring-blue-300
-              "
-              onClick={() => {
-                setPage(page + 1);
-              }}
-            >
-              Next
-            </button>
-          </div>
-          <RegisterUserModal
-            setShowModal={setShowModal}
-            showModal={showModal}
-            onRegister={handleRegisterUser}
-          />
+
+          <RegisterUserModal setShowModal={setShowModal} showModal={showModal} onRegister={handleRegisterUser} />
           {selectedUser && (
             <UpdateUserModal
               setShowModal={setShowUpdateModal}
@@ -228,15 +185,7 @@ export const Users: FunctionComponent = () => {
               onUpdate={handleUpdateUser}
             />
           )}
-          {/* Toast de confirmación */}
-          {showToast && (
-            <Toast
-              message="Usuario registrado exitosamente"
-              onClose={() => {
-                setShowToast(false);
-              }}
-            />
-          )}
+          {showToast && <Toast message="Usuario registrado exitosamente" onClose={() => { setShowToast(false); }} />}
         </main>
       </div>
     </Layout>
